trailersplus/requirements.txt
aiohttp==3.7.3
amqp==2.6.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
git+https://github.com/j2devcore/sdk-python.git@createdocexc#egg=authorizenet
beautifulsoup4==4.8.2
billiard==3.6.3.0
boto3==1.14.63
botocore==1.17.63
celery==4.4.7
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
distlib==0.3.2
Django==3.0.11
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-cogwheels==0.3
django-cors-headers==3.5.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.0
django-filter==2.4.0
django-modelcluster==5.1
django-modeltranslation==0.16.1
django-redis==4.12.1
django-storages==1.10.1
django-taggit==1.3.0
django-timezone-field==4.1.1
django-treebeard==4.3.1
djangorestframework==3.11.2
docutils==0.15.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
filelock==3.0.12
flower==0.9.5
geoip2==4.1.0
geopy==2.3.0
gunicorn==20.0.4
html5lib==1.1
humanize==3.1.0
idna==2.10
jdcal==1.4.1
jmespath==0.10.0
kombu==4.6.11
l18n==2020.6.1
maxminddb==2.0.3
multidict==5.1.0
openpyxl==3.0.5
Pillow==7.2.0
pipenv==2021.5.29
prometheus-client==0.8.0
psycopg2-binary==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-memcached==1.59
pytz==2020.4
PyXB==1.2.5
redis==3.5.3
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==3.0.0
tornado==6.1
tqdm==4.15.0
trustpilot==9.0.0
typing-extensions==3.7.4.3
ujson==4.0.1
Unidecode==1.1.1
urllib3==1.25.11
vine==1.3.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
wagtail==2.10
wagtail-cache==1.0.1
wagtail-inventory==1.2
wagtail-meta-preview==1.0.3
wagtail-modeltranslation==0.10.15
sentry-sdk==1.3.1
wagtailmedia==0.7.0
wagtailmenus==3.0.2
wagtailuiplus==1.3.14
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
yarl==1.6.3
